test(typecheck): isolate slot binding regression
